+++ src/main/java/com/Core/LangRepository.java
@@ -9,10 +9,10 @@ class LangRepository {
 
    LangRepository() {
        languages=new ArrayList<>();
-       languages.add(new Lang(1L,"Hello", "en"));
-       languages.add(new Lang(2L,"joł", "pl"));
+       languages.add(new Lang(1,"Hello", "en"));
+       languages.add(new Lang(2,"joł", "pl"));
     }
-    Optional<Lang> findById  (Long id){
+    Optional<Lang> findById  (Integer id){
        return languages
                .stream()
                .filter(lang -> lang.getId().equals(id))
